Removed vvo2.py
